+++ src/main/java/kg/example/levantee/model/enums/order/OrderStatusConverter.java
@@ -4,16 +4,16 @@ import jakarta.persistence.AttributeConverter;
 import jakarta.persistence.Converter;
 
 @Converter
-public class OrderStatusConverter implements AttributeConverter<OrderStatus, Integer> {
+public class OrderStatusConverter implements AttributeConverter<OrderStatus, Short> {
 
     @Override
-    public Integer convertToDatabaseColumn(OrderStatus status) {
+    public Short convertToDatabaseColumn(OrderStatus status) {
         if (status == null) return null;
         return status.id;
     }
 
     @Override
-    public OrderStatus convertToEntityAttribute(Integer dbValue) {
+    public OrderStatus convertToEntityAttribute(Short dbValue) {
         if (dbValue == null) return null;
         return OrderStatus.fromId(dbValue);
     }
